Refactor admin and demo functions and import method
